complete transition from test.go to emulator
add contractor and originator functions
move CLI interface to ofcli/, run a server only instance at test.go
