+++ rpc/entities.go
@@ -6,12 +6,15 @@ import (
 	"net/http"
 
 	solar "github.com/OpenFinancing/openfinancing/platforms/solar"
+	utils "github.com/OpenFinancing/openfinancing/utils"
 )
 
 func setupEntityRPCs() {
 	validateEntity()
 	getPreOriginatedContracts()
 	getOriginatedContracts()
+	getProposedContracts()
+	addCollateral()
 }
 
 func EntityValidateHelper(w http.ResponseWriter, r *http.Request) (solar.Entity, error) {
@@ -78,5 +81,50 @@ func getOriginatedContracts() {
 			return
 		}
 		MarshalSend(w, r, x)
+	})
+}
+
+func getProposedContracts() {
+	http.HandleFunc("/entity/getproposed", func(w http.ResponseWriter, r *http.Request) {
+		checkGet(w, r)
+		prepEntity, err := EntityValidateHelper(w, r)
+		if err != nil {
+			errorHandler(w, r, http.StatusNotFound)
+			return
+		}
+
+		x, err := solar.RetrieveContractorProjects(solar.ProposedProject, prepEntity.U.Index)
+		if err != nil {
+			errorHandler(w, r, http.StatusNotFound)
+			return
+		}
+		MarshalSend(w, r, x)
+	})
+}
+
+func addCollateral() {
+	//func (contractor *Entity) AddCollateral(amount float64, data string) error {
+	http.HandleFunc("/entity/addcollateral", func(w http.ResponseWriter, r *http.Request) {
+		checkGet(w, r)
+		prepEntity, err := EntityValidateHelper(w, r)
+		if err != nil || r.URL.Query()["amount"] == nil || r.URL.Query()["collateral"] == nil {
+			errorHandler(w, r, http.StatusNotFound)
+			return
+		}
+
+		collateralAmount, err := utils.StoFWithCheck(r.URL.Query()["amount"][0])
+		if err != nil {
+			errorHandler(w, r, http.StatusNotFound)
+			return
+		}
+
+		collateralData := r.URL.Query()["collateral"][0]
+		err = prepEntity.AddCollateral(collateralAmount, collateralData)
+		if err != nil {
+			errorHandler(w, r, http.StatusNotFound)
+			return
+		}
+
+		Send200(w, r)
 	})
 }
